split reprojected export out into own test file
